// src/types/ErrorType.ts
export type propsType = {
  name?: string
  metaTitle?: string
  slug?: string
  description?: string
  Categories?: string
  imageFile?: File | null
  heading?: string
  primaryButtonText?: string
  primaryButtonLink?: string
  productCode?: string
  variantName?: string
  variantSize?: string
  variantColor?: string
  variantProductSKU?: string
  variantQuantity?: string
  variantRegularPrice?: string
  variantSalePrice?: string
  nameUser?: string
  emailAddressUser?: string
  phoneUser?: string
  oldPassword?: string
  password?: string
  newPassword?: string
  confirmPassword?: string
}

export type Errors = {
  name?: string
  metaTitle?: string
  slug?: string
  description?: string
  Categories?: string
  imageFile?: string
  heading?: string
  primaryButtonText?: string
  primaryButtonLink?: string
  productCode?: string
  variantName?: string
  variantSize?: string
  variantColor?: string
  variantProductSKU?: string
  variantQuantity?: string
  variantRegularPrice?: string
  variantSalePrice?: string
  nameUser?: string
  emailAddressUser?: string
  phoneUser?: string
  oldPassword?: string
  password?: string
  newPassword?: string
  confirmPassword?: string
}